Prevent Cleanup from causing deployment to fail
Having many Cleanup tasks run in parallel causes ls to return with an error
due to disappearing files. I've changed the cleanup job to run ls on the
filespec it's looking for, rather than running ls on the parent directory and
then scanning through the results in our code.

I've also changed Cleanup so that if ls fails it's now non-fatal. The Cleanup
job will stop, but the deployment will continue. This is so that automated
builds (i.e. iCAS building the entire platform) do not fail.