Fixes for generator stage creation during package deployment
Add package deployment stages with a generic name, and add the list of hosts
to the description afterwards. Otherwise extra stages end up being created
when attempting to group packages on different hosts within the same stage.

Also bugfix for icas dbmigrations: Need to maintain a list of dbmigrations that
have been added in order to end up with a list of non-prefixed packages at the
end.